Rename test classes to unify package structure and improve clarity
